# docs/swagger/swagger.yaml
basePath: /
definitions:
  cmd_api_apis.AddCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  cmd_api_apis.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  cmd_api_apis.LoginRequest:
    properties:
      email:
        maxLength: 320
        type: string
      password:
        maxLength: 72
        type: string
    required:
    - email
    - password
    type: object
  cmd_api_apis.ProductRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      name:
        maxLength: 200
        type: string
      price:
        minimum: 0
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/models.ProductStatus'
        enum:
        - active
        - inactive
      stock:
        minimum: 0
        type: integer
    required:
    - name
    type: object
  cmd_api_apis.RegisterRequest:
    properties:
      email:
        maxLength: 320
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  cmd_api_apis.UpdateCartItemRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
  cmd_api_apis.UserResponse:
    properties:
      created_at:
        format: date-time
        type: string
      email:
        example: customer@example.com
        type: string
      id:
        format: uuid
        type: string
      name:
        example: Customer
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - customer
        - admin
    type: object
  github_com_linenxing_e-commerce-system_base_response.ErrorBody:
    properties:
      code:
        example: invalid_request
        type: string
      message:
        example: request data is invalid
        type: string
    type: object
  github_com_linenxing_e-commerce-system_services_auth.AuthOutput:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/models.UserResp'
    type: object
  models.CartItemResp:
    properties:
      id:
        type: string
      product:
        $ref: '#/definitions/models.ProductResp'
      quantity:
        type: integer
      subtotal:
        type: integer
    type: object
  models.CartResp:
    properties:
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.CartItemResp'
        type: array
      total_price:
        type: integer
    type: object
  models.OrderItemResp:
    properties:
      price:
        type: integer
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
    type: object
  models.OrderResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.OrderItemResp'
        type: array
      status:
        $ref: '#/definitions/models.OrderStatus'
      total_price:
        type: integer
    type: object
  models.OrderStatus:
    enum:
    - completed
    type: string
    x-enum-varnames:
    - OrderStatusCompleted
  models.ProductResp:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        format: uuid
        type: string
      name:
        type: string
      price:
        example: 1000
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/models.ProductStatus'
        enum:
        - active
        - inactive
      stock:
        example: 10
        type: integer
      updated_at:
        type: string
    type: object
  models.ProductStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - ProductStatusActive
    - ProductStatusInactive
  models.UserResp:
    properties:
      created_at:
        format: date-time
        type: string
      email:
        example: customer@example.com
        type: string
      id:
        format: uuid
        type: string
      name:
        example: Customer
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - customer
        - admin
    type: object
  models.UserRole:
    enum:
    - customer
    - admin
    type: string
    x-enum-varnames:
    - UserRoleCustomer
    - UserRoleAdmin
info:
  contact: {}
  description: Backend APIs for the simple e-commerce MVP.
  title: Simple E-Commerce API
  version: "1.0"
paths:
  /admin/products:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductResp'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Admin Products
  /admin/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Disable product
      tags:
      - Admin Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductResp'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Admin Products
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password and return a JWT access token.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_services_auth.AuthOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
      summary: Login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a customer account and return a JWT access token.
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_services_auth.AuthOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
      summary: Register customer
      tags:
      - Authentication
  /cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CartResp'
      security:
      - BearerAuth: []
      summary: Get cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Cart item
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.AddCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CartResp'
      security:
      - BearerAuth: []
      summary: Add cart item
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remove cart item
      tags:
      - Cart
    put:
      consumes:
      - application/json
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      - description: Quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api_apis.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CartResp'
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - Cart
  /health:
    get:
      description: Return the current API health status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api_apis.HealthResponse'
      summary: Health check
      tags:
      - System
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OrderResp'
            type: array
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - Orders
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OrderResp'
      security:
      - BearerAuth: []
      summary: Create order from cart
      tags:
      - Orders
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderResp'
      security:
      - BearerAuth: []
      summary: Get my order
      tags:
      - Orders
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductResp'
            type: array
      summary: List products
      tags:
      - Products
  /products/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductResp'
      summary: Get product
      tags:
      - Products
  /users/me:
    get:
      description: Return the user represented by the bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api_apis.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_linenxing_e-commerce-system_base_response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the Bearer prefix, for example: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"